Added 'eos' signal that can stop the create loop'
Added 'eos' signal and corresponding callback in order to be able to break from the create loop; due to how the plugin handles the creation of the buffer when the trigger is enabled, the plugin gets stuck in a loop and can never change states. If you emit an 'eos' signal in the GstAravis element, however, a flag will be set that will break the loop with an GST_FLOW_EOS return code, which will itself put the aravis source plugin in a PAUSED state.